feat(x64): complete 64-bit timestamp migration cascade (server side)
Resolve the full TimeT64 ripple to 0 warnings across the server:
- Widen the remaining timestamp-carrying APIs: CHARACTER skill-next-read
  / quest-flag setters+getters and m_aiPremiumTimes; quest PC SaveFlag;
  CPrivManager Give*/RequestGive* duration/end_time params + SPriv*Data
  end_time fields; marriage Add + TMarriage marry_time; TItemAward
  dwSocket0-2 (reads the now-BIGINT item_award.socketN columns).
- Replace the int-only MINMAX clamp of privilege end_time (an absolute
  epoch) with a 64-bit clamp -- the old static_cast<int> clamp was itself
  a 2038 truncation.
- Fix Cache.cpp socket compare buffer/width (was sizeof(long), now matches
  TimeT64 alSockets); DB INSERT socketN format %u -> %lld; lua_pushnumber
  of a flag value cast to lua_Number; socket-as-vnum cast for GetTable.
- affect.h/questpc.h/priv_manager.h/marriage.h/ItemAwardManager.h include
  common/length.h for TimeT64.

Server x64: 0 compile errors, 0 warnings (/W3). Timestamp migration done
on the server+shared+DB side; client edits + protocol/DB lockstep deploy
remain for the client port.

+++ Lead-Server-Source/game/src/DragonSoul.cpp
@@ -1031,7 +1031,7 @@ int DSManager::LeftTime(LPITEM pItem) const
 	// For now timer based on wear Only the dragon soul stone does not disappear even if time runs out. .
 	if (pItem->GetProto()->cLimitTimerBasedOnWearIndex >= 0)
 	{
-		return pItem->GetSocket(ITEM_SOCKET_REMAIN_SEC);
+		return static_cast<int>(pItem->GetSocket(ITEM_SOCKET_REMAIN_SEC));
 	}
 	// different limit type Since all Dragon Soul Stones disappear over time, the items entered here are judged to have some time left. .
 	else
